# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnalystGPT Enterprise",
    "description": "Enterprise-grade analytics platform providing data upload, cleaning, quality assessment, analytics, reporting, AI-powered business insights, Power BI integration, REST APIs, and Streamlit dashboard visualisation.",
    "contact": {
      "name": "Amal Jose"
    },
    "license": {
      "name": "MIT"
    },
    "version": "12.0.0"
  },
  "paths": {
...
    "/api/governance/preview": {
      "post": {
        "tags": [
          "Governance",
          "Governance"
        ],
        "summary": "Non-Destructive Cleaning Preview",
        "description": "Simulates cleaning transformations and calculates quality metrics without persisting artifacts or creating pipeline run records.",
        "operationId": "preview_cleaning_api_governance_preview_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-User-Name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Name"
            }
          },
          {
            "name": "X-User-Email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleaningPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleaningPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/governance/dataset-versions": {
      "get": {
        "tags": [
          "Governance",
          "Governance"
        ],
        "summary": "List Dataset Versions",
        "operationId": "list_dataset_versions_api_governance_dataset_versions_get",
        "parameters": [
          {
            "name": "is_source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Source"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-User-Name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Name"
            }
          },
          {
            "name": "X-User-Email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatasetVersionResponse"
                  },
                  "title": "Response List Dataset Versions Api Governance Dataset Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/governance/dataset-versions/{version_id}": {
      "get": {
        "tags": [
          "Governance",
          "Governance"
        ],
        "summary": "Get Dataset Version",
        "operationId": "get_dataset_version_api_governance_dataset_versions__version_id__get",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-User-Name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Name"
            }
          },
          {
            "name": "X-User-Email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/governance/lineage/run/{pipeline_run_id}": {
      "get": {
        "tags": [
          "Governance",
          "Governance"
        ],
        "summary": "Get Pipeline Run Lineage",
        "operationId": "get_pipeline_run_lineage_api_governance_lineage_run__pipeline_run_id__get",
        "parameters": [
          {
            "name": "pipeline_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pipeline Run Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-User-Name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Name"
            }
          },
          {
            "name": "X-User-Email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CleaningConfigRequest": {
        "properties": {
          "missing_value_policy": {
            "type": "string",
            "title": "Missing Value Policy",
            "description": "Named missing value strategy (e.g. DROP_ROWS, FILL_NUMERIC_MEAN, CUSTOM_POLICY)",
            "default": "DROP_ROWS"
          },
          "null_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Null Threshold",
            "description": "Null percentage threshold for DROP_COLUMNS_ABOVE_THRESHOLD"
          },
          "fill_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fill Value",
            "description": "Constant fill value for FILL_CATEGORICAL_CONSTANT"
          },
          "affected_columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Columns",
            "description": "Specific columns to target. If null, applies across all eligible columns."
          },
          "custom_strategy_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Strategy Name",
            "description": "Name of registered custom strategy handler if policy is CUSTOM_POLICY"
          },
          "custom_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Params",
            "description": "Validated parameters dictionary for custom strategy"
          }
        },
        "type": "object",
        "title": "CleaningConfigRequest",
        "description": "Request model for creating/configuring cleaning policies."
      },
...
      "CleaningPreviewRequest": {
        "properties": {
          "dataset_version_id": {
            "type": "string",
            "title": "Dataset Version Id",
            "description": "UUID of registered source dataset version to preview"
          },
          "cleaning_config": {
            "$ref": "#/components/schemas/CleaningConfigRequest",
            "description": "Proposed cleaning configuration to test"
          },
          "sample_rows": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Sample Rows",
            "description": "Number of sample rows to return in before/after preview comparison",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "dataset_version_id"
        ],
        "title": "CleaningPreviewRequest",
        "description": "Request model for non-destructive cleaning preview.\nRequires a registered immutable dataset_version_id reference."
      },
      "CleaningPreviewResponse": {
        "properties": {
          "quality_comparison": {
            "$ref": "#/components/schemas/QualityComparisonResponse"
          },
          "columns_removed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns Removed"
          },
          "affected_columns_detail": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Affected Columns Detail"
          },
          "preview_sample_source": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Preview Sample Source"
          },
          "preview_sample_cleaned": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Preview Sample Cleaned"
          }
        },
        "type": "object",
        "required": [
          "quality_comparison",
          "columns_removed",
          "affected_columns_detail",
          "preview_sample_source",
          "preview_sample_cleaned"
        ],
        "title": "CleaningPreviewResponse",
        "description": "Response model returned by non-destructive preview endpoint."
      },
...
      "DatasetVersionResponse": {
        "properties": {
          "version_id": {
            "type": "string",
            "title": "Version Id"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "byte_size": {
            "type": "integer",
            "title": "Byte Size"
          },
          "checksum_sha256": {
            "type": "string",
            "title": "Checksum Sha256"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "column_count": {
            "type": "integer",
            "title": "Column Count"
          },
          "is_source": {
            "type": "boolean",
            "title": "Is Source"
          },
          "storage_path": {
            "type": "string",
            "title": "Storage Path"
          },
          "parent_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Version Id"
          },
          "dataset_schema_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Schema Json"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "version_id",
          "source_filename",
          "content_type",
          "byte_size",
          "checksum_sha256",
          "row_count",
          "column_count",
          "is_source",
          "storage_path"
        ],
        "title": "DatasetVersionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LineageResponse": {
        "properties": {
          "source_dataset_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "cleaned_dataset_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "cleaning_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CleaningConfigResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "cleaning_execution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CleaningExecutionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "LineageResponse"
      },
...
      "QualityComparisonResponse": {
        "properties": {
          "rows_removed": {
            "type": "integer",
            "title": "Rows Removed"
          },
          "pct_rows_removed": {
            "type": "number",
            "title": "Pct Rows Removed"
          },
          "missing_values_resolved": {
            "type": "integer",
            "title": "Missing Values Resolved"
          },
          "completeness_gain": {
            "type": "number",
            "title": "Completeness Gain"
          },
          "has_material_loss": {
            "type": "boolean",
            "title": "Has Material Loss"
          },
          "source_completeness_pct": {
            "type": "number",
            "title": "Source Completeness Pct"
          },
          "cleaned_completeness_pct": {
            "type": "number",
            "title": "Cleaned Completeness Pct"
          }
        },
        "type": "object",
        "required": [
          "rows_removed",
          "pct_rows_removed",
          "missing_values_resolved",
          "completeness_gain",
          "has_material_loss",
          "source_completeness_pct",
          "cleaned_completeness_pct"
        ],
        "title": "QualityComparisonResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}